Put all lib files in ino directory
