--- Software/Arduino/include/motor_control.hpp
@@ -1,10 +1,0 @@
-#include <AccelStepper.h>
-
-class motorControl :: public AccelStepper {
-public:
-    motorControl(stepPin, dirPin);
-    void actuateMotors(long speed);
-
-private:
-    AccelStepper stepper;
-};
