+++ Software/Arduino/main/motor_control.cpp
@@ -1,6 +1,6 @@
-#include "../include/motor_control.hpp"
+#include "motor_control.hpp"
 
-motorControl::motorControl(AccelStepper::DRIVER, stepPin, dirPin) : stepper(AccelStepper::DRIVER, stepPin, dirPin) {
+motorControl::motorControl(int stepPin, int dirPin) : stepper(AccelStepper::DRIVER, stepPin, dirPin) {
     stepper.disableOutputs();
     stepper.setMaxSpeed(10000);
     stepper.setCurrentPosition(0);
@@ -8,7 +8,7 @@ motorControl::motorControl(AccelStepper::DRIVER, stepPin, dirPin) : stepper(Acce
 }
 
 void motorControl::actuateMotors(long speed) {
-  float stepperSpeed = degrees * 1023.0 / 270;
+  float stepperSpeed = speed * 1023.0 / 270;
   stepperSpeed = constrain(stepperSpeed, 0, 1023);
   stepper.setSpeed(stepperSpeed);
   stepper.runSpeed();
